Switch renaming of APKs off for debuggable builds
There seem to be issues with the new AGP when the APK name changes between builds which forces us to do a clean build every now and then. If we don't rename the APK this should no longer be the case and we don't need the full information in debug builds anyways.